# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "auth",
        "description": "API documentation for my Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dnd/auth/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/info": {
            "get": {
                "tags": [
                    "Аунтификация"
                ],
                "summary": "Информация об пользователе",
                "operationId": "e02a030195280c5d3b90ef29e4af697c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|pVEhLK1z9QCN6CSn5FG5djNqiWC6XZc1zoeDfMLE375c7750"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "test"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@test.ru"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Authentication Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/check": {
            "get": {
                "tags": [
                    "Аунтификация"
                ],
                "summary": "Проверка авторизации пользователя",
                "operationId": "d2ed4b1e31da6a05545c5b32aed9e8dc",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 1|pVEhLK1z9QCN6CSn5FG5djNqiWC6XZc1zoeDfMLE375c7750"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "authenticated": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AuthenticationErrorResponse": {
                "title": "AuthenticationErrorResponse",
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            }
        }
    },
...
}